+++ internal/pkg/runtime/engine/singularity/process_linux.go
@@ -41,6 +41,22 @@ import (
 	"github.com/hpcng/singularity/pkg/sylog"
 	"github.com/hpcng/singularity/pkg/util/rlimit"
 	specs "github.com/opencontainers/runtime-spec/specs-go"
+<<<<<<< HEAD
+=======
+	"github.com/sylabs/singularity/internal/pkg/instance"
+	"github.com/sylabs/singularity/internal/pkg/plugin"
+	"github.com/sylabs/singularity/internal/pkg/security"
+	"github.com/sylabs/singularity/internal/pkg/util/env"
+	"github.com/sylabs/singularity/internal/pkg/util/fs/files"
+	"github.com/sylabs/singularity/internal/pkg/util/machine"
+	"github.com/sylabs/singularity/internal/pkg/util/shell"
+	"github.com/sylabs/singularity/internal/pkg/util/shell/interpreter"
+	"github.com/sylabs/singularity/internal/pkg/util/user"
+	singularitycallback "github.com/sylabs/singularity/pkg/plugin/callback/runtime/engine/singularity"
+	singularityConfig "github.com/sylabs/singularity/pkg/runtime/engine/singularity/config"
+	"github.com/sylabs/singularity/pkg/sylog"
+	"github.com/sylabs/singularity/pkg/util/rlimit"
+>>>>>>> 724490cc2... fix: Allow env vars with literal $ via \\\$ in SINGULARITYENV_
 	"golang.org/x/crypto/ssh/terminal"
 	"golang.org/x/sys/unix"
 	"mvdan.cc/sh/v3/interp"
@@ -631,18 +647,22 @@ func injectEnvHandler(senv map[string]string) interpreter.OpenHandler {
 			`
 			b.WriteString(fmt.Sprintf(defaultPathSnippet, env.DefaultPath))
 
+			// https://github.com/sylabs/singularity/issues/43
+			// We wrap the value of the export in double quotes manually, and do not use
+			// go's %q format string as it prevents passing an escaped literal $ in
+			// a SINGULARITYENV_ as \$
 			snippet := `
 			if test -v %[1]s; then
 				sylog debug "Overriding %[1]s environment variable"
 			fi
-			export %[1]s=%[2]q
+			export %[1]s="%[2]s"
 			`
 			for key, value := range senv {
 				if key == "LD_LIBRARY_PATH" && value != "" {
 					b.WriteString(fmt.Sprintf(snippet, key, value+":/.singularity.d/libs"))
 					continue
 				}
-				b.WriteString(fmt.Sprintf(snippet, key, value))
+				b.WriteString(fmt.Sprintf(snippet, key, shell.EscapeQuotes(value)))
 			}
 		})