+++ CHANGELOG.md
@@ -17,6 +17,15 @@ _With the release of `v3.0.0`, we're introducing a new changelog format in an at
 
 _The old changelog can be found in the `release-2.6` branch_
 
+<<<<<<< HEAD
+=======
+### Bug Fixes
+
+  - Allow escaped `\$` in a SINGULARITYENV_ var to set a literal `$` in
+    a container env var.
+
+## v3.8.0 [2021-05-26]
+>>>>>>> 724490cc2... fix: Allow env vars with literal $ via \\\$ in SINGULARITYENV_
 
 # v3.8.0 - [2021-06-15]